+++ internal/store/store.go
@@ -105,6 +105,7 @@ type Store struct {
 	imapUpdates chan imapBackend.Update
 
 	isSyncRunning bool
+	syncCooldown  cooldown
 	addressMode   addressMode
 }
 
@@ -148,6 +149,9 @@ func New(
 		lock:          &sync.RWMutex{},
 		log:           l,
 	}
+
+	// Minimal increase is event pollInterval, doubles every failed retry up to 5 minutes.
+	store.syncCooldown.setExponentialWait(pollInterval, 2, 5*time.Minute)
 
 	if err = store.init(firstInit); err != nil {
 		l.WithError(err).Error("Could not initialise store, attempting to close")